requirements.txt
matplotlib==3.4.2
Pillow==8.1.0
streamlit==0.74.1
torch==1.10.0
torchvision==0.11.1